requirements.txt
keyboard
requests
Pillow
python-dotenv
win10toast
